Remove landing-page supervisor logs
