+++ landing-page/src/data/observer.ts
@@ -1,0 +1,3 @@
+export async function getObserverLogs() {
+  return fetch("/observer/logs");
+}
